+++ src/main/java/subway/line/application/LineService.java
@@ -4,8 +4,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import subway.line.Line;
 import subway.line.infra.LineRepository;
-import subway.line.controller.LineRequest;
-import subway.line.controller.LineResponse;
+import subway.line.presentation.LineRequest;
+import subway.line.presentation.LineResponse;
 import subway.line.exception.LineNotFoundException;
 import subway.station.Station;
 import subway.station.application.StationService;
