+++ src/main/java/subway/line/presentation/LineRequest.java
@@ -1,4 +1,4 @@
-package subway.line.controller;
+package subway.line.presentation;
 
 public class LineRequest {
