+++ src/test/java/subway/line/LineRestAssured.java
@@ -4,7 +4,7 @@ import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import org.springframework.http.MediaType;
-import subway.line.controller.LineRequest;
+import subway.line.presentation.LineRequest;
 
 public class LineRestAssured {
